Added ability to set paths
These paths are (semi) smart and will navigate around obsticles in their way. Currently if a distraction is set the NPC will move towards it but never return to the default path. If a distraction is NOT set the NPC will path forever.

Working on a way to make it so if the NPC hits its head against a while for long enough it automatically jiggles itself out of being stuck